study_tracker/requirements.txt
Flask
Flask-MySQLdb
Flask-APScheduler
